% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforced_VS.R
\name{modelPredict_VS}
\alias{modelPredict_VS}
\title{Risk prediction on the test set, variable selection version}
\usage{
modelPredict_VS(list_fpcaFit, list_cvfit, Xtest, Ztest, startT, tau)
}
\arguments{
\item{list_fpcaFit}{Obtained FPCA decomposition from \code{modelFit_VS}.}

\item{list_cvfit}{Obtained elastic net logistic regression from \code{modelFit_VS}.}

\item{Xtest}{Longitudinal biomarker data for subjects in the test set, matrix of \eqn{testn} by \eqn{nTotal}. Missing values
are denoted by NA.}

\item{Ztest}{Other baseline covariates for subjects in the test set.}

\item{startT}{Time of the first prediction, denoted by \eqn{t_1} in the manuscript. For instance, if the time grids are \eqn{{0,1/60,2/60,...,1}},
then startT = 25 means that the first prediction is made at \eqn{t = 24/60}.}

\item{tau}{The threshold value \eqn{\tau}.}
}
\value{
\item{final.label}{Predicted outcome \eqn{Y} for subjects in the test set, vector of length \eqn{testn}.}
\item{avg.cost}{Average cost when we applied this prediction procedure to the test set. }
\item{cost}{Cost for each subject, vector of length \eqn{testn}. For some subjects, we make a definite decision early. For others, we follow up with a long period of time. Hence the cost is different for each individual. }
}
\description{
For a fixed threshold value \eqn{\tau}, \code{modelPredict_VS} predicts the outcome \eqn{Y}  for subjects in the test set.
This function also outputs the cost associated with the prediction procedure. This function is used when the baseline
covariates Z are high-dimensional.
}
\examples{
library(reinforcedPred)
set.seed(1)

# take the example training data (high dimensional Z) from the reinforcedPred package
# see documentation for details about the data set train_data_mulZ
Y <- as.numeric(train_data_mulZ$Y)
tildeX.missing <- as.matrix(train_data_mulZ[,2:62])
Z <- as.matrix(train_data_mulZ[,63:dim(train_data_mulZ)[2]])

# analysis starts
budget <- 45
folds <- 5
startT <- 25

result <- reinforced_VS(Y, tildeX.missing, Z, budget, folds, startT, pve = 0.99, nbasis = 10)

# obtained elastic net logistic regression fit and FPCA decompositions
list_cvfit <- (result$final.result)$list_cvfit
list_fpcaFit <- (result$final.result)$list_fpcaFit

# optimal tau that minimizes the misclassification error under the budget constraint
final.tau <- result$final.tau
final.tau

# use the fitted model to predict the label Y for subjects in the test data
# see documentation for details about the data set test_data_mulZ
testY <- as.numeric(test_data_mulZ$testY)
test.tildeX.missing <- as.matrix(test_data_mulZ[,2:62])
test.Z <- as.matrix(test_data_mulZ[,63:dim(test_data_mulZ)[2]])

pred <- modelPredict_VS(list_fpcaFit, list_cvfit, test.tildeX.missing, test.Z, startT, final.tau)

# predicted outcome Y for each subject in the test data
predY.test <- pred$final.label
# misclassification error
mis.error <- sum(predY.test != testY, na.rm = TRUE) / sum(!is.na(testY))
mis.error

# the average cost when we applied the prediction procedure to the test data
pred$avg.cost
}
