% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforced_VS.R
\docType{data}
\name{train_data_mulZ}
\alias{train_data_mulZ}
\title{Example training data (high dimensional Z)}
\format{A data frame with 400 rows and 112 columns. The 1st column is the outcome variable Y, starting from the 2nd
column to 62nd column is the longitudinal biomarker at 61 time grids, the 63rd column to 112nd column are other baseline covariate Z.}
\source{
A simulated data set
}
\usage{
train_data_mulZ
}
\description{
Example training data (high dimensional Z)
}
\keyword{datasets}
